refactor part2 + number and json strategies
