+++ admin-frontend/app_singleapp/lib/widgets/features/feature_dashboard_constants.dart
@@ -3,4 +3,4 @@ final unselectedRowHeightPerStrategy = 40.0;
 final selectedRowHeight = 260.0;
 final selectedRowHeightPerStrategy = 40.0;
 final headerHeight = 40.0;
-final cellWidth = 300.0;
+final cellWidth = 350.0;
