+++ admin-frontend/app_singleapp/lib/widgets/features/feature_value_cell.dart
@@ -1,10 +1,10 @@
 import 'package:app_singleapp/widgets/features/feature_dashboard_constants.dart';
-import 'package:app_singleapp/widgets/features/feature_value_row_json.dart';
-import 'package:app_singleapp/widgets/features/feature_value_row_number.dart';
-import 'package:app_singleapp/widgets/features/feature_value_row_string.dart';
 import 'package:app_singleapp/widgets/features/table-collapsed-view/non_boolean_value_cell.dart';
 import 'package:app_singleapp/widgets/features/table-collapsed-view/value_boolean_cell.dart';
-import 'package:app_singleapp/widgets/features/table-expanded-view/boolean_cell_holder.dart';
+import 'package:app_singleapp/widgets/features/table-expanded-view/boolean/boolean_cell_holder.dart';
+import 'package:app_singleapp/widgets/features/table-expanded-view/json/json_cell_holder.dart';
+import 'package:app_singleapp/widgets/features/table-expanded-view/number/number_cell_holder.dart';
+import 'package:app_singleapp/widgets/features/table-expanded-view/string/string_cell_holder.dart';
 import 'package:app_singleapp/widgets/features/tabs_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
@@ -68,25 +68,25 @@ class FeatureValueCell extends StatelessWidget {
             final fvBloc = tabsBloc.featureValueBlocs[feature.key];
             switch (feature.valueType) {
               case FeatureValueType.BOOLEAN:
-                cellWidget = FeatureValueBooleanCellEditor(
+                cellWidget = BooleanCellHolder(
                     environmentFeatureValue: efv,
                     feature: feature,
                     fvBloc: fvBloc);
                 break;
               case FeatureValueType.STRING:
-                cellWidget = FeatureValueStringCellEditor(
+                cellWidget = StringCellHolder(
                     environmentFeatureValue: efv,
                     feature: feature,
                     fvBloc: fvBloc);
                 break;
               case FeatureValueType.NUMBER:
-                cellWidget = FeatureValueNumberCellEditor(
+                cellWidget = NumberCellHolder(
                     environmentFeatureValue: efv,
                     feature: feature,
                     fvBloc: fvBloc);
                 break;
               case FeatureValueType.JSON:
-                cellWidget = FeatureValueJsonCellEditor(
+                cellWidget = JsonCellHolder(
                     environmentFeatureValue: efv,
                     feature: feature,
                     fvBloc: fvBloc);
