+++ admin-frontend/app_singleapp/lib/widgets/features/table-expanded-view/boolean/edit_boolean_value_dropdown_widget.dart
