+++ admin-frontend/app_singleapp/lib/widgets/features/table-expanded-view/delete_strategy_icon_button.dart
