+++ admin-frontend/app_singleapp/lib/widgets/features/table-expanded-view/edit_value_strategy_link_button.dart
@@ -1,7 +1,7 @@
 import 'package:app_singleapp/widgets/common/fh_underline_button.dart';
-import 'package:app_singleapp/widgets/features/create-strategy-widget.dart';
 import 'package:app_singleapp/widgets/features/custom_strategy_bloc.dart';
 import 'package:app_singleapp/widgets/features/per_feature_state_tracking_bloc.dart';
+import 'package:app_singleapp/widgets/features/table-expanded-view/create-strategy-widget.dart';
 import 'package:flutter/material.dart';
 import 'package:mrapi/api.dart';
