+++ admin-frontend/app_singleapp/lib/widgets/features/table-expanded-view/split_rollout_button.dart
@@ -1,6 +1,6 @@
-import 'package:app_singleapp/widgets/features/create-strategy-widget.dart';
 import 'package:app_singleapp/widgets/features/custom_strategy_bloc.dart';
 import 'package:app_singleapp/widgets/features/per_feature_state_tracking_bloc.dart';
+import 'package:app_singleapp/widgets/features/table-expanded-view/create-strategy-widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_icons/flutter_icons.dart';
