+++ admin-frontend/app_singleapp/lib/widgets/features/table-expanded-view/strategy_card_widget.dart
@@ -1,18 +1,18 @@
 import 'package:app_singleapp/widgets/features/custom_strategy_bloc.dart';
-import 'package:app_singleapp/widgets/features/delete_strategy_icon_button.dart';
 import 'package:app_singleapp/widgets/features/per_feature_state_tracking_bloc.dart';
+import 'package:app_singleapp/widgets/features/table-expanded-view/delete_strategy_icon_button.dart';
 import 'package:app_singleapp/widgets/features/table-expanded-view/edit_value_strategy_link_button.dart';
 import 'package:flutter/material.dart';
 import 'package:mrapi/api.dart';
 
-class StrategyCardsListWidget extends StatelessWidget {
+class StrategyCardWidget extends StatelessWidget {
   final bool editable;
   final Widget editableHolderWidget;
   final RolloutStrategy rolloutStrategy;
   final PerFeatureStateTrackingBloc fvBloc;
   final CustomStrategyBloc strBloc;
 
-  const StrategyCardsListWidget(
+  const StrategyCardWidget(
       {Key key,
       @required this.editable,
       @required this.editableHolderWidget,
@@ -35,7 +35,7 @@ class StrategyCardsListWidget extends StatelessWidget {
               mainAxisSize: MainAxisSize.min,
               children: <Widget>[
                 Expanded(
-                    flex: 5,
+                    flex: 4,
                     child: rolloutStrategy == null
                         ? Text('default',
                             style: Theme.of(context).textTheme.caption)
@@ -45,9 +45,9 @@ class StrategyCardsListWidget extends StatelessWidget {
                             fvBloc: fvBloc,
                             strBloc: strBloc,
                           )),
-                Expanded(flex: 3, child: editableHolderWidget),
+                Expanded(flex: 5, child: editableHolderWidget),
                 Expanded(
-                  flex: 2,
+                  flex: 1,
                   child: rolloutStrategy != null
                       ? DeleteStrategyIconButton(
                           editable: editable, rolloutStrategy: rolloutStrategy, strBloc: strBloc,)
